Fit a natural effect model

This function is similar to R function neModel from the

This function is similar to R function neModel from the medflex package.

neModelDS(formula, family, expData, se, nBoot, newobj)

Arguments

formula
a formula object providing a symbolic description of the natural effect model.

family
aa description of the error distribution and link function to be used in the model. For glm this can be a character string naming a family function, a family function or the result of a call to a family function. For glm.fit only the third option is supported.

expData
the expanded dataset (of class "expData").

se
character string indicating the type of standard errors to be calculated. The default type is based on the bootstrap.

nBoot
number of bootstrap replicates.

newobj
a character string that provides the name for the output object that is stored on the data servers. Default neModel.out.

Value

a summary table of the object of class 'neModel'.

Details

The function 'neModelDS' is used to fit a natural effect model on the expanded dataset.

Author

Demetris Avraam, for DataSHIELD Development Team